requirements.txt
apiview==1.3.25
asgiref==3.3.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
constantly==15.1.0
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.5
django-extensions==3.1.3
django-import-export==2.5.0
django-rest-framework==0.1.0
django-socket-server==0.0.4
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.1
image==1.5.33
incremental==21.3.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.0
psycopg2==2.9.1
pycparser==2.20
pydot==1.4.2
pydotplus==2.0.2
pygame==2.0.1
pyparsing==2.4.7
pytz==2020.5
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
